+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Referencing an object by name in a script file now returns the latest when there are collisions e.g. "[ExtractableCohort]" would return the latest one (created during the script execution session)
 - Bump YamlDotNet from 11.2.0 to 11.2.1
 - Bump SecurityCodeScan.VS2019 from 5.1.0 to 5.2.1
+- Command 'Set' now shows as Impossible for property 'ID'
 
 ## Fixed